Update Crosstool-ng 1.26.0 to 1.27.0 on linux-arm64 (keep same GCC, glibc etc...)
Update OpenSSL from 1.1.1o to 1.1.1w
Update Ninja from 1.11.1 to 1.12.1
Update git from 2.42.0 to 2.48.1
Update Liquidpromt from 1.11 to 1.12 and add to env ARG

+++ imagefiles/install-liquidprompt-binary.sh
@@ -7,12 +7,17 @@ if ! command -v git &> /dev/null; then
 	exit 1
 fi
 
+if [[ -z "${LIQUIDPROMT_VERSION}" ]]; then
+  echo >&2 'error: LIQUIDPROMT_VERSION env. variable must be set to a non-empty value'
+  exit 1
+fi
+
 cd /usr/share
 
 # Remove liquidprompt if exists
 rm -rf liquidprompt
 
-git clone "https://github.com/nojhan/liquidprompt.git" -b v_1.11
+git clone "https://github.com/nojhan/liquidprompt.git" --depth 1 -b "${LIQUIDPROMT_VERSION}" liquidprompt
 
 cat << EOF >> /root/.bashrc
